+++ CMakeLists.txt
@@ -17,8 +17,8 @@ include_directories(${CMAKE_CURRENT_SOURCE_DIR}/src)
 set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR})
 
 add_library(rk_solvers
-            include/runge_kutta.h
-            src/runge_kutta.cpp
+        include/Solver_base.h
+        src/Solver_base.cpp
             include/solver_runge_kutta_3.h
             src/solver_runge_kutta_3.cpp
             include/solver_runge_kutta_4.h
