+++ include/solver_runge_kutta_4.h
@@ -6,7 +6,7 @@
 #include "phy.h"
 #include <format>
 #include <tuple>
-#include "runge_kutta.h"
+#include "Solver_base.h"
 
 namespace rk4 {
