+++ include/solver_runge_kutta_adaptive.h
@@ -3,7 +3,7 @@
 
 #include "solver_runge_kutta_4.h"
 #include "solver_runge_kutta_5.h"
-#include "runge_kutta.h"
+#include "Solver_base.h"
 #include "phy.h"
 
 namespace rk_add {
